tg-bot/requirements.txt
aiogram
openpyxl
pandas
python-decouple
icecream
apscheduler
aiohttp-socks
asyncpg-lite
requests
prettytable